+++ src/test/java/org/selenium/demo/ChromeTest.java
@@ -1,47 +1,13 @@
 package org.selenium.demo;
 
-import org.junit.*;
+import org.junit.Assert;
+import org.junit.Ignore;
+import org.junit.Test;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
+import org.selenium.demo.utils.ChromeDriverInit;
 
-import java.io.InputStream;
-import java.util.Properties;
-
-public class ChromeTest {
-
-    private WebDriver driver;
-
-    @Before
-    public void setUp(){
-        Properties properties = initProperties();
-        System.setProperty("webdriver.chrome.driver", properties.getProperty("chrome.driver"));
-        driver = initDriver();
-    }
-
-    @After
-    public void tearDown() throws Exception {
-        if (driver != null) {
-            driver.close();
-            driver.quit();
-        }
-    }
-
-    private WebDriver initDriver() {
-        return new ChromeDriver();
-    }
-
-    private Properties initProperties() {
-        Properties properties = new Properties();
-        try {
-            InputStream fileInputStream = getClass().getClassLoader().getResourceAsStream("pom.properties");
-            properties.load(fileInputStream);
-        } catch (Exception e) {
-            System.err.println(e.getMessage());
-            throw new RuntimeException(e.getMessage());
-        }
-        return properties;
-    }
+public class ChromeTest extends ChromeDriverInit {
 
     /** The following code is for the Chrome Driver
      You also need to download the ChromeDriver executable
